observability/grafana/tools/dashboard-saver/requirements.txt
GitPython==3.1.32
GitPython==3.1.36
Requests==2.31.0
